Now shows all resources as bucket urls; new tasks push this to bucket
All the resources are on the bucket. But WIP for testing the new deletion code, and some changes will still need to be done. Marked the things I know of in new todo in the code. Replaced the unlink and file testing. Need to still work with real_path in some code.

The new tasks to sync files work