Uses estimates plugin to adjust budget; shows estimates (1.0.30)
Listens to estimate changes, if the estimate is linked to the bid will sum up all estimates associated with the bid to be the new budget.

Shows the estimate table in the show bids, above the files.

If estimate plugin is not used, then still works, will not do the above